Split OpenAI Responses engine helpers
