+++ pkg/steps/ai/openai_responses/engine.go
@@ -12,7 +12,6 @@ import (
 
 	"github.com/go-go-golems/geppetto/pkg/events"
 	"github.com/go-go-golems/geppetto/pkg/inference/engine"
-	"github.com/go-go-golems/geppetto/pkg/inference/tools"
 	geppettoobs "github.com/go-go-golems/geppetto/pkg/observability"
 	"github.com/go-go-golems/geppetto/pkg/security"
 	"github.com/go-go-golems/geppetto/pkg/steps/ai/runtimeattrib"
@@ -30,13 +29,6 @@ type Engine struct {
 	settings            *settings.InferenceSettings
 	observer            geppettoobs.Observer
 	observabilityConfig geppettoobs.Config
-}
-
-type usageTotals struct {
-	inputTokens     int
-	outputTokens    int
-	cachedTokens    int
-	reasoningTokens int
 }
 
 func NewEngine(s *settings.InferenceSettings, opts ...EngineOption) (*Engine, error) {
@@ -1108,176 +1100,4 @@ func (e *Engine) RunInference(ctx context.Context, t *turns.Turn) (*turns.Turn, 
 	}
 	e.publishEvent(ctx, events.NewFinalEvent(metadata, message))
 	return t, nil
-}
-
-func (e *Engine) attachToolsToResponsesRequest(ctx context.Context, t *turns.Turn, reqBody *responsesRequest) error {
-	if t == nil || reqBody == nil {
-		return nil
-	}
-
-	engineTools := tools.AdvertisedToolDefinitionsFromContext(ctx)
-
-	var toolCfg engine.ToolConfig
-	if cfg, ok, err := engine.KeyToolConfig.Get(t.Data); err != nil {
-		return errors.Wrap(err, "get tool config")
-	} else if ok {
-		toolCfg = cfg
-	}
-
-	if len(engineTools) > 0 && toolCfg.Enabled {
-		converted, err := e.PrepareToolsForResponses(engineTools, toolCfg)
-		if err != nil {
-			return err
-		}
-		if arr, ok := converted.([]any); ok && len(arr) > 0 {
-			reqBody.Tools = arr
-			// Responses API: omit tool_choice for function tools to allow model selection
-			reqBody.ToolChoice = nil
-			if toolCfg.MaxParallelTools > 1 {
-				b := true
-				reqBody.ParallelToolCalls = &b
-			} else if toolCfg.MaxParallelTools == 1 {
-				b := false
-				reqBody.ParallelToolCalls = &b
-			}
-			log.Debug().Int("tool_count", len(arr)).Interface("tool_choice", reqBody.ToolChoice).Msg("Responses: tools attached to request")
-		}
-	}
-
-	if builtins, ok, err := turns.KeyResponsesServerTools.Get(t.Data); err != nil {
-		return errors.Wrap(err, "get responses server tools")
-	} else if ok && len(builtins) > 0 {
-		reqBody.Tools = append(reqBody.Tools, builtins...)
-		log.Debug().Int("builtin_tool_count", len(builtins)).Msg("Responses: server-side tools attached to request")
-	}
-
-	return nil
-}
-
-func mustMarshalJSON(v any) []byte {
-	b, err := json.Marshal(v)
-	if err != nil {
-		return []byte("{}")
-	}
-	return b
-}
-
-func parseUsageTotalsFromEnvelope(envelope map[string]any) (usageTotals, bool) {
-	if envelope == nil {
-		return usageTotals{}, false
-	}
-	usage, ok := envelope["usage"].(map[string]any)
-	if !ok {
-		if respObj, hasResponse := envelope["response"].(map[string]any); hasResponse {
-			usage, ok = respObj["usage"].(map[string]any)
-		}
-	}
-	if !ok || usage == nil {
-		return usageTotals{}, false
-	}
-	return parseUsageTotals(usage), true
-}
-
-func parseUsageTotalsFromResponse(resp responsesResponse) (usageTotals, bool) {
-	if totals, ok := parseUsageTotalsFromRawUsage(resp.Usage); ok {
-		return totals, true
-	}
-	if resp.Response != nil {
-		if totals, ok := parseUsageTotalsFromRawUsage(resp.Response.Usage); ok {
-			return totals, true
-		}
-	}
-	return usageTotals{}, false
-}
-
-func parseUsageTotalsFromRawUsage(raw json.RawMessage) (usageTotals, bool) {
-	if len(raw) == 0 || string(raw) == "null" {
-		return usageTotals{}, false
-	}
-	var usage map[string]any
-	if err := json.Unmarshal(raw, &usage); err != nil {
-		return usageTotals{}, false
-	}
-	return parseUsageTotals(usage), true
-}
-
-func parseUsageTotals(usage map[string]any) usageTotals {
-	ret := usageTotals{}
-	if v, ok := toInt(usage["input_tokens"]); ok {
-		ret.inputTokens = v
-	}
-	if v, ok := toInt(usage["output_tokens"]); ok {
-		ret.outputTokens = v
-	}
-	if inputDetails, ok := usage["input_tokens_details"].(map[string]any); ok {
-		if v, ok := toInt(inputDetails["cached_tokens"]); ok {
-			ret.cachedTokens = v
-		}
-	} else if v, ok := toInt(usage["cached_tokens"]); ok {
-		ret.cachedTokens = v
-	}
-	if outputDetails, ok := usage["output_tokens_details"].(map[string]any); ok {
-		if v, ok := toInt(outputDetails["reasoning_tokens"]); ok {
-			ret.reasoningTokens = v
-		}
-	} else if v, ok := toInt(usage["reasoning_tokens"]); ok {
-		ret.reasoningTokens = v
-	}
-	return ret
-}
-
-func normalizeResponsesEventName(eventName string) string {
-	switch eventName {
-	case "response.reasoning.delta":
-		return "response.reasoning_text.delta"
-	case "response.reasoning.done":
-		return "response.reasoning_text.done"
-	default:
-		return eventName
-	}
-}
-
-func toInt(v any) (int, bool) {
-	const maxInt = int(^uint(0) >> 1)
-	const minInt = -maxInt - 1
-
-	switch x := v.(type) {
-	case float64:
-		if x > float64(maxInt) || x < float64(minInt) {
-			return 0, false
-		}
-		return int(x), true
-	case float32:
-		f := float64(x)
-		if f > float64(maxInt) || f < float64(minInt) {
-			return 0, false
-		}
-		return int(x), true
-	case int:
-		return x, true
-	case int32:
-		return int(x), true
-	case int64:
-		if x > int64(maxInt) || x < int64(minInt) {
-			return 0, false
-		}
-		return int(x), true
-	case uint:
-		if uint64(x) > uint64(maxInt) {
-			return 0, false
-		}
-		return int(x), true
-	case uint32:
-		if uint64(x) > uint64(maxInt) {
-			return 0, false
-		}
-		return int(x), true
-	case uint64:
-		if x > uint64(maxInt) {
-			return 0, false
-		}
-		return int(x), true
-	default:
-		return 0, false
-	}
 }
